+++ src/pages/BookmarkPage.js
@@ -1,5 +1,5 @@
 import { useSelector } from "react-redux";
-import SubPageTemplate from "../components/SubPageTemplate";
+import SubPageTemplate from "./template/SubPageTemplate";
 
 function BookmarkPage() {
   const bookmarkList = useSelector((state) => state.bookmarkList);
