+++ src/pages/ProductPage.js
@@ -1,5 +1,5 @@
 import { useSelector } from "react-redux";
-import SubPageTemplate from "../components/SubPageTemplate";
+import SubPageTemplate from "./template/SubPageTemplate";
 
 function ProductPage() {
   const productList = useSelector((state) => state.productList);
